Merged PR 4088: low level pipeline concepts
# [DRAFT]

![image (2).png](https://dev.azure.com/immunomics/8edd37b2-5b67-4c9e-a5f3-77a69004b53d/_apis/git/repositories/9c3987ab-eb30-450e-bf61-df476177b394/pullRequests/4088/attachments/image%20%282%29.png)
These are the top level classes. Create an instance of `MlPipelineConfig` to define the nodes and their dependencies. Nodes need to implement the `IExecutable` interface. Pass the `MlPipelineConfig` to an `MlPipelineProvider` class instance and follow the rest of the example in `piplines3._example._di_container.py` to see how the pipeline is executed. The `Pipeline3ExampleApplication` class has the call that executes the pipeline.