Merged PR 4194: pipeline ux concepts
Made some clearer distinctions between the clients for `building`, `representing`, and `executing` a pipeline. Each of the three sections have a few components for implementing the various capabilities and I hope the new structure is a little easier to follow. I took the unfinished storage code out of the repo and will add it back in a followup PR.

- use the builder components to add nodes, dependencies, and executables to a pipeline
- use the builder to get a pipeline components object
- use the pipeline components to get a session object
- use the session object to execute the pipeline