Merged PR 4725: move display_dag to oneml
`display_dag` is calling existing `workflow_to_svg` to build the SVG representation of a workflow, and then calling IPython specific functions to display that SVG in a jupyter notebook.
Because it needs to import from IPython, it was defined in immunocam-ml.  But now that we also have an immunopipeline kernel, that doesn't make sense.
Instead, I moved it to oneml-processors, and imported from IPython within the function.  This means the function would fail if called in an environment that does not have IPython installed.