Merged PR 4879: remote pipeline execution
Initial implementation of remote execution in kubernetes. We added a builder method to wrap an executable in remote logic and implemented a version of the remote class that executes in k8s. There are some basics being added around controlling settings and getting to the currently active session and node classes so that we can implement classes that are initialized before the session. Added a few examples of simple data publishing that does not require the node to be specified and created some bare bones blob code for storing/retrieving results along with serialization libraries.