Merged PR 5003: make remote executables serializable
Main thing added here was the `SessionComponents` class that has `add_component()` and `get_component()` methods. The example in my demo app uses these methods to set the di container as a component in the session and shows an example of a pipeline step reaching into the di container for a fake spark session.

The example pipeline can be run like this from the oneml-app component:
```
immunocli-next ci docker-build --image-tag lolo-prototype --push && \
    democli run-pipeline --image immunodata.azurecr.io/oneml-app:lolo-prototype
```