Merged PR 5780: Persist trained eval pipeline
Persist eval pipeline feature working e2e.

There are a few capabilities missing from oneml-pipelines io system to enable coding this by assigning publishers to ports.  I implemented using processors that write and read, leveraging IReadData and IWriteData within these processors.  This can be refactored once we add the missing capabilities.

The fitted parameters are persisted using a `WriteToNodeBasedUriProcessor` which uses session and node ids to build the uri it will write to (so very similar to `IPublishPipelineData`).  Crucially, it also outputs that uri.  The uris for all fitted parameters are consumed as inputs by `CreateFittedEvalPipelineProcessor` node, that hard-codes them as inputs to `ReadFromUriProcessor`s in the fitted-eval-pipeline it builds and outputs.  That fitted-eval-pipeline is persisted using another `WriteToNodeBasedUriProcessor`, and also outputted by the full pipeline.  All uris are also outputted by the full pipeline.

Giving `CreateFittedEvalPipelineProcessor` access to the uris to which the fitted parameters were written is one of the capabilities missing from current oneml-pipelines io system.

Another missing capability is a plugin system for mapping types and store scheme (memory, file, blob, etc) to writers/readers.
This system needs to enable:
* Registering a reader/writer per type and scheme.  In oneml-processors ports are associated with type annotations, not type ids, so I think simply using the type as a key in a registry will not work.  At the minimum we will need a default behaviour per scheme.
* Providing a type and getting a mapping from scheme to reader/writer.  This is needed so we don't couple the time we know the type with the time we know the scheme.