Merged PR 5796: Allow processor outputs to be defined using NamedTuple instead of TypedDict
Dill does not handle deserializing TypedDict class objects.
This means we can't use it to deserialize pipelines with processors that were defined in notebooks, if those processors use TypedDict to define their outputs.
The problem only applies when processors are defined in notebooks, b/c otherwise the class objects are not serialized.

This PR allows using NamedTuples instead.
In the future, we should probably use only NamedTuples, for consistency.