Merged PR 6044: clean io concepts
I kept all the changes in an `io2` package so it's easier to compare but some of the classes are unmodified from `io`.

# Local Venv
we've been rolling out a change in habitats to make all venvs be generated locally by poetry. the `poetry.toml` files add this setting and it might require you to run `poetry env remove 3.10` and then `poetry install` again. sometimes this change gets confused by direnv but changing up a directory and coming back will resolve it. this change improves things by making it easier for us to find this venv and seems to improve how both vscode and pycharm find this env automatically.

in this pr I made sure this file exists in all our components so you might need to recreate your venv if you run into errors.

# Plugins

added a draft of a definition of a new plugin concept that represents the ability for others to hook into pipeline behaviors through lifecycle events and add functionality to the system. we now have multiple flavors of plugins like `OnemlAppPlugin` and `OnemlIoPlugin` and we have the beginning of a pattern for future plugin definitions.

# Pipeline Data

added a new way of thinking about io to a new `oneml.io2` package. i'm leaving the old one intact because i did not fully replace the functionality but i want to take a slice and refine it. hoping to follow this pr with some new ones to tackle the remaining io rough edges (doesn't have to be right away).

# Core Entities
I took out most of the `*Context` classes and replaced them with more direct uses of the core entities below:

- `PipelineSession`
- `PipelineNode`
- `PipelinePort`

there are now contexts that map directly to these types so you can depend on `ContextProvider[PipelineNode]` and treat it as a simple callable that returns that data structure.

i also added a context within the io layer that is opened when we are attempting to persist a specific node port.

# DI

i removed a few old unused classes like `PrivateOnemlAppServices` and started following a pattern of putting these service id mapping classes in the same file as the related di container that implements their providers. there is a one-to-many relationship from a library class to its ServiceIds and there is a one-to-one relationship between a ServiceId and the container method that marks it as its ServiceProvider. Each ServiceId is a concrete use-case for a certain class (service).