create rats-apps component (#76)
i've started with making everything under a `rats.apps` package with a
full test suite to show some examples of the apis. this time i made sure
to expose the public api properly.

- added a `ConfigId` class that just refers to `ServiceId` providers
that return a `namedtuple` of any kind. our expectation is that these
entries are serializable data structures.
- added group namespaces to allow for simpler implementations of
different kinds of containers
- `get()` and `get_group()` both look for fallback definitions

---------

Co-authored-by: jzazo <SC-vf957@microsoft.com>
Co-authored-by: Javier Zazo <javierzazo@microsoft.com>