feature: devtools app (#116)
this is the first attempt at using devtools to manage the file watcher
logic and run formatting tools automatically. this allows us to control
the list of watchers using the container groups functionality and allows
us to expose custom messaging to people working in pycharm. i'm fairly
confident this is possible in vscode but want to experiment here before
trying to adopt any patterns.