✨ feat(/workspaces/rats/rats-devtools/src/python/rats/command_tree): Added CommandTree concept to simplify writing highly nested CLI commands
This adds a CommandTree object that can be converted into click commands (and can be easily adapted to handle other frameworks as well)). It re-implements most of the argparse specific functionality of smple-parsing, but using click and introduces an intermediate representation + pattern that allows generalization to other frameworks. It's a mostly unopinionated structure that encourages users to follow CLI UX best practices, while abstracting away the complexity of the underlying framework. Things have been designed to allow power users to directly interact with the underlying framework to avoid the abstraction becoming a limitation.