♻️  refactor(rats-devtools/src/python/rats): Convert all commands to CommandTree
This uses the structure in Lorenzo's PR [#157](https://github.com/microsoft/rats/pull/157) to validate the CommandTree pattern. These additions help DX, and enable us to invert the dependency between the commands and their subcommands, while enabling control over whether a given command is open for expansion, or restricted to expansion by the command's maintainer. I've replicated all of the functionality used by the devtools component except for the click.Path suppo. That should be straightforward to enable by adding an explicit click type to the ClickConversion type annotation.