More devtools focus (#272)
trying to quickly untangle any aspects of the devtools needed to run
things in other projects. with the changes in this pr, i can run the
`ci` commands on any component that happens to use the same linting and
testing tools. trying to improve docs i started with deleting a lot of
old code and a couple unused layers of abstraction around the cli tools.
i hope i did most of the changes without breaking compatibility, but
bumping versions just in case.