more rats.apps apis (#411)
# lots of documentation work
still trying to make the onboarding experience better and adding lots of
examples across the code. added a lot of docstrings to
the main `rats.apps` package and tried to make as many things as
possible individually executable. i also took another pass at updating
the `mkdocs.yaml` file with settings that make the api docs more
readable. cross-refs and other navigation should be
much cleaner now.

needs a lot more work but i had to stop somewhere.

# removed `rats.apps.SimpleApplication`
the `SimpleApplication` class was adding a lot of magic that was often
removed as soon as the examples became a tiny bit more involved. i think
the new structure allows more a more natural progression from simple to
complex and doesn't introduce
a lot of concepts and service ids that have to immediately be changed. i
think i tracked down any old uses and transitioned to
the new style. the devtools commands are the best example of this right
now.

# new apis for creating and running apps

- `apps.AppContainer` is a service container with `execute()`
- `apps.AppPlugin` is a `Callable[[Container], AppContainer]`
- `apps.ContainerPlugin` is a `Callable[[Container], Container]`
- `apps.PluginMixin` adds the default constructor to turn a class into a
valid app or container plugin
- `apps.run(*apps)` runs a list of app containers
- `apps.run_plugin(*cls)` takes a list of `apps.AppPlugin` and runs them
- `apps.AppBundle` helps us create app instances with provided plugins
- `apps.static_group()/static_service()` help in the creation of static
containers
- `apps.Runtime` interface removed `execute_callable()` method
- `apps.StandardRuntime` taken out of the simple apps file and kept

# transitioned to shallower, simpler apps
initially we were creating a small set of applications that had many
plugin containers loaded into them. but this is too difficult
to maintain when our plugin containers have little in common. with the
new apis above, it should be easier to create many
small application classes that we execute inside a larger one, so there
is much less shared state.

as an example, i turned the logging plugin into a new application that
we just run individually before other things. this should
make the logging app trivial to run and maintain.

# initial `rats_e2e.apps` package
started adding a new package called `rats_e2e` that is meant to ship
with the wheel and contains end-to-end tests for parts of
the libraries. i tried documenting these tests heavily so they can both
test capabilities but also serve as good examples for
users.

this new set of tests is hopefully enough to remove the need for the two
example components i had made in the past. this should be much simpler
to maintain.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>