Feature/plugin containers and code removal (#449)
renaming the `apps.PluginContainers` class to
`apps.PythonEntryPointContainer` for clarity since we have a few other
things called plugins already. also using this pr to delete as much code
outside of `rats-apps` as possible. leaving the components in place but
will be deleting the src/test contents and removing the components from
being built and released. we think these are possible future components
but want to clear the way to release rats-apps and rats packages without
broken dependencies.

---------

Co-authored-by: Lorenzo Pisani <lorenzo.pisani@microsoft.com>