github releases from ci (#485)
this change adds a step to the release ci workflow that generates a
github release draft that can be published when we tag a release. it
only generates this release on main branch builds and tags, but the main
branch builds are marked as `--prerelease`. since they are drafts,
they're only visible to contributors until one of us publishes it, at
least for now. i think in the future we could make tagged releases
publish since that matches the behavior for publishing wheels.

you can see a draft release generated by this pr before disabling pr
releases here:
https://github.com/microsoft/rats/releases/tag/untagged-1f584fa16e62c912c920


![image](https://github.com/user-attachments/assets/2b2c30b7-783f-4e9b-a696-66102705c095)

refs #119