fix container group caching issues (#531)
calling `rats.apps.Container.get_group()` a second time for a given
service id was returning an empty list, because we're caching the group
generator, which gets exhausted after the first time looping through
results. fixed by caching the list instead of the generator. also added
a unit test for this behavior.