rats.projects improvements and support for deeper component directory nesting (#554)
mainly addressing #546 while also cleaning up a lot of logic, and adding
a lot of easy tests. we are limiting the components to being up to 4
directories deep within the project root, so that we minimize
performance impact of searching the entire code base.

also fixes coverage settings in devtools component (#553)