fix docs rough edges (#492) and remove deprecated docs code (#556)
- the `rats-docs mkdocs-serve` and `rats-docs mkdocs-build` commands are
now removed and replaced with `rats-docs serve` and `rats-docs build`
respectively.
- our `rats-docs` command does almost nothing now, removing the last
remaining rough edges described in #492 and instead using symlinks to
unify the docs across components. leaving the command in the code
because of a possible future where we can generate the scaffolding for
people trying to first set up documentation in their repo.
- updated docs about docs :)