lighter dependencies in devtools (#598)
removed the `twine` dependency because it's not longer used, and moved
all the documentation related dependencies to the `dev` group because we
want the end user to select the set of packages on their own based on
what they decide to put in their `mkdocs.yaml` config. this does mean
that the `rats-docs` commands will fail if the user hasn't added the
`mkdocs` package as a dependency.